Move scheduling to separate file
