+++ .gitignore
@@ -1,2 +1,2 @@
-./__pycache__
-.venv
+venv
+__pycache__/**
