+++ r2.txt
@@ -1,0 +1,18 @@
+annotated-types==0.6.0
+APScheduler==3.10.4
+asttokens==2.4.1
+colorama==0.4.6
+executing==2.0.1
+halo==0.0.31
+icecream==2.1.3
+log-symbols==0.0.14
+pydantic==2.5.2
+pydantic_core==2.14.5
+Pygments==2.17.2
+pytz==2023.3.post1
+six==1.16.0
+spinners==0.0.24
+termcolor==2.3.0
+typing_extensions==4.8.0
+tzlocal==5.2
+watchdog==3.0.0
